feat: Add more guide on how to do camera calibration
Camera calibration is usually needed in order to flatten out image to true 2D/3D representation. Reason for needing to do this is because when doing taking photos with cameras, we will run into curved lens which would distort the image slightly. This is done by having the camera take several chessboard image and resize it accordingly.